refactoring sdk and example
