+++ .gitignore
@@ -71,3 +71,4 @@ android/keystores/debug.keystore
 
 # generated by bob
 lib/
+dist/
