+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "flourish-sdk-web",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "npx rollup -c"
