+++ src/config.js
@@ -1,18 +1,15 @@
-import ProdConfig from './config.production';
-import StagingConfig from './config.staging';
-
 let BACKEND_API_URL = new Map();
 let FRONTEND_URL = new Map();
 let FLOURISH_SDK_APP_VERSION = new Map();
 
-BACKEND_API_URL.set('staging', StagingConfig.BACKEND_API_URL);
-BACKEND_API_URL.set('production', ProdConfig.BACKEND_API_URL);
+BACKEND_API_URL.set('staging', 'https://api-stg.flourishfi.com/api/v1');
+BACKEND_API_URL.set('production', 'https://api.flourishfi.com/api/v1');
 
-FRONTEND_URL.set('staging', StagingConfig.FRONTEND_URL);
-FRONTEND_URL.set('production', ProdConfig.FRONTEND_URL);
+FRONTEND_URL.set('staging', 'https://platform-stg.flourishfi.com');
+FRONTEND_URL.set('production', 'https://platform.flourishfi.com');
 
-FLOURISH_SDK_APP_VERSION.set('staging', StagingConfig.FLOURISH_SDK_APP_VERSION);
-FLOURISH_SDK_APP_VERSION.set('production', ProdConfig.FLOURISH_SDK_APP_VERSION);
+FLOURISH_SDK_APP_VERSION.set('staging', '1.0.0');
+FLOURISH_SDK_APP_VERSION.set('production', '1.0.0');
 
 const Config = {
   BACKEND_API_URL,
