--- src/config/config.production.js
@@ -1,6 +1,0 @@
-export default {
-    FRONTEND_URL: 'https://platform.flourishfi.com',
-    BACKEND_API_URL: 'https://api.flourishfi.com/api/v1',
-    FLOURISH_SDK_APP_VERSION: '1.0.0'
-  };
-
